+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [2.3.0] - 2020-09-10
+
+Pypi: https://pypi.org/project/cobib/2.3.0/
+
 ### Added
 - Logging functionality has been added. The verbosity level can be controlled via `-v` (INFO) and `-vv` (DEBUG).
   As soon as the TUI starts, all logging output is redirected to `/tmp/cobib.log`.
@@ -237,7 +242,8 @@ Note: this release was not marked MAJOR because this is still a WIP and early
 - initial version with a basic `sqlite3`-based database
 
 
-[Unreleased]: https://gitlab.com/mrossinek/cobib/-/compare/v2.2.2...master
+[Unreleased]: https://gitlab.com/mrossinek/cobib/-/compare/v2.3.0...master
+[2.3.0]: https://gitlab.com/mrossinek/cobib/-/tags/v2.3.0
 [2.2.2]: https://gitlab.com/mrossinek/cobib/-/tags/v2.2.2
 [2.2.1]: https://gitlab.com/mrossinek/cobib/-/tags/v2.2.1
 [2.2.0]: https://gitlab.com/mrossinek/cobib/-/tags/v2.2.0